// src/shared/types.ts
export enum SortOption {
  AlphabeticalFromAtoZ = "alphabeticalFromAtoZ",
  AlphabeticalFromZtoA = "alphabeticalFromZtoA",
  TimeAddedNewFirst = "timeAddedNewFirst",
  TimeAddedNewLast = "timeAddedNewLast",
  AmountAscending = "amountAscending",
  AmountDescending = "amountDescending",
  CompletedFirst = "completedFirst",
  CompletedLast = "completedLast",
}